refactor: split accounts actions into thin wrappers and impl modules
Move strategy, stakeholders, contacts, matching, deals, CRUD, and onepager logic out of the accounts God-File while keeping the public server-action API stable.
