+++ app/dashboard/accounts/actions.ts
@@ -1,132 +1,78 @@
 'use server'
 
-import { revalidatePath } from 'next/cache'
-import { ROUTES } from '@/lib/routes'
-import { createServerSupabaseClient } from '@/lib/supabase/server'
-import { asJson, asTableInsert, asTableUpdate } from '@/lib/supabase/db-types'
-import * as XLSX from 'xlsx'
-import {
-  normalizeCompanyAccountStatus,
-  type CompanyAccountStatusValue,
-} from '@/lib/accounts/company-account-status'
 import type { PartnerCategory } from '@/lib/accounts/company-entity'
+import type {
+  AccountDealRow,
+  CompanyAccountStatusValue,
+  CompanyRefRow,
+  CompanyStrategyRow,
+  ContactPersonRow,
+  DealSignalRow,
+  RecommendedReference,
+  RoadmapProjectRow,
+  StakeholderRole,
+  StakeholderRow,
+} from './account-action-types'
 import {
-  parseCompaniesImportRow,
-} from '@/lib/accounts/companies-import-parse'
-import { enrichBulkImportRowFromBrandfetch } from '@/lib/accounts/resolve-company-for-import'
-import { ensureBrandfetchDarkLogoUrl } from '@/lib/brandfetch/logo-theme-url'
-import { formatIndustryDisplay, resolveIndustryId } from '@/lib/constants/industries'
+  getCompanyStrategyImpl,
+  upsertCompanyStrategyImpl,
+  getRoadmapProjectsImpl,
+  upsertRoadmapProjectImpl,
+  deleteRoadmapProjectImpl,
+} from './strategy-roadmap-impl'
 import {
-  discoverAndSaveCompanyNewsrooms,
-  scheduleCompanyNewsroomDiscovery,
-} from '@/lib/market-signals/discover-company-newsroom'
-import { parseProfileRoles } from '@/lib/roles/profile-roles'
-import { profileIsSalesRestricted } from '@/lib/roles/profile-guards'
+  getStakeholdersImpl,
+  createStakeholderImpl,
+  updateStakeholderImpl,
+  deleteStakeholderImpl,
+} from './stakeholders-impl'
+import {
+  updateExternalContactBuyingCenterRoleImpl,
+  getContactsByCompanyIdImpl,
+  setCompanyInternalReferenceApprovalContactImpl,
+  createContactPersonImpl,
+  updateContactPersonImpl,
+  deleteContactPersonImpl,
+} from './contacts-impl'
+import {
+  getRecommendedReferencesImpl,
+  getReferencesForOrgImpl,
+  getReferencesByCompanyIdImpl,
+  getRecommendedReferencesForAccountImpl,
+} from './account-match-impl'
+import {
+  getActiveDealsByCompanyIdImpl,
+  getExpiringDealsByCompanyIdImpl,
+} from './account-deals-impl'
+import {
+  updateCompanyAccountStatusImpl,
+  toggleCompanyFavoriteImpl,
+  createCompanyImpl,
+  createPartnerImpl,
+  bulkCreateCompaniesFromSheetImpl,
+  updateCompanyImpl,
+  deleteCompanyWithDataImpl,
+} from './company-crud-impl'
+import { generateOnePagerHtmlImpl } from './onepager-impl'
 
-export type { CompanyAccountStatusValue } from '@/lib/accounts/company-account-status'
-
-export type CompanyStrategyRow = {
-  id: string
-  company_id: string
-  company_goals: string | null
-  red_flags: string | null
-  competition: string | null
-  next_steps: string | null
-  value_proposition?: string | null
-  metrics_pain?: string | null
-  mh_assessment?: Record<string, unknown> | null
-  updated_at: string | null
-}
-
-export type { StakeholderRole } from '@/lib/accounts/stakeholder-role'
-import type { StakeholderRole } from '@/lib/accounts/stakeholder-role'
-
-export type StakeholderRow = {
-  id: string
-  company_id: string
-  name: string
-  title: string | null
-  role: StakeholderRole
-  influence_level?: string | null
-  attitude?: string | null
-  notes?: string | null
-  linkedin_url?: string | null
-  priorities_topics?: string | null
-  last_contact_at?: string | null
-  last_interaction_at?: string | null
-  sentiment?: string | null
-  created_at: string
-  updated_at: string | null
-}
-
-export type CompanyRefRow = {
-  id: string
-  title: string
-  status: string
-  project_status: string | null
-  industry?: string | null
-  country?: string | null
-  created_at: string
-}
-
-/** Opportunity Roadmap: ein Projekt pro Zeile */
-export type RoadmapProjectRow = {
-  id: string
-  company_id: string
-  project_name: string
-  estimated_value: string | null
-  status: string | null
-  target_date: string | null
-  tags: string | null
-  created_at: string
-  updated_at: string | null
-}
-
-/** Empfohlene Referenz für Smart-Matching inkl. Score 0–100 und Begründung */
-export type RecommendedReference = {
-  id: string
-  title: string
-  company_name: string | null
-  matchType: 'industry_and_tags' | 'tags_only' | 'industry_only'
-  /** Score 0–100 (Branche 50, Themen 30, Größe/Region 20) */
-  score: number
-  /** Für Tooltip: Warum dieses Match? */
-  matchReasons: { industry: boolean; tags: boolean; sizeRegion: boolean }
-}
+export type {
+  CompanyAccountStatusValue,
+  CompanyStrategyRow,
+  StakeholderRole,
+  StakeholderRow,
+  CompanyRefRow,
+  RoadmapProjectRow,
+  RecommendedReference,
+  ContactPersonRow,
+  ExternalContactRow,
+  AccountDealRow,
+  DealSignalRow,
+} from './account-action-types'
 
 export async function getCompanyStrategy(
   companyId: string
 ): Promise<CompanyStrategyRow | null> {
-  const supabase = await createServerSupabaseClient()
-  const full = await supabase
-    .from('company_strategies')
-    // DB-Spalten: main_goals, competitive_situation; wir mappen per Alias auf unsere Feldnamen
-    .select(
-      'id, company_id, company_goals:main_goals, red_flags, competition:competitive_situation, next_steps, value_proposition, metrics_pain, mh_assessment, updated_at'
-    )
-    .eq('company_id', companyId)
-    .maybeSingle()
-  if (!full.error) return full.data as CompanyStrategyRow | null
-
-  const msg = (full.error.message ?? '').toLowerCase()
-  const missingOptionalColumns =
-    msg.includes('metrics_pain') || msg.includes('mh_assessment') || msg.includes('value_proposition')
-  if (!missingOptionalColumns) return null
-
-  const fallback = await supabase
-    .from('company_strategies')
-    .select(
-      'id, company_id, company_goals:main_goals, red_flags, competition:competitive_situation, next_steps, updated_at'
-    )
-    .eq('company_id', companyId)
-    .maybeSingle()
-  if (fallback.error || !fallback.data) return null
-  return {
-    ...(fallback.data as CompanyStrategyRow),
-    value_proposition: null,
-    metrics_pain: null,
-    mh_assessment: null,
-  }
+  return getCompanyStrategyImpl(companyId)
 }
 
 export async function upsertCompanyStrategy(
@@ -141,58 +87,11 @@ export async function upsertCompanyStrategy(
     mh_assessment?: Record<string, unknown> | null
   }
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const full = await supabase.from('company_strategies').upsert(
-    asTableInsert<'company_strategies'>({
-      company_id: companyId,
-      // DB-Spalten heißen main_goals und competitive_situation
-      main_goals: payload.company_goals ?? null,
-      red_flags: payload.red_flags ?? null,
-      competitive_situation: payload.competition ?? null,
-      next_steps: payload.next_steps ?? null,
-      value_proposition: payload.value_proposition ?? null,
-      metrics_pain: payload.metrics_pain ?? null,
-      mh_assessment: asJson(payload.mh_assessment ?? {}),
-      updated_at: new Date().toISOString(),
-    }),
-    { onConflict: 'company_id' }
-  )
-  if (!full.error) {
-    revalidatePath(ROUTES.accounts)
-    revalidatePath(ROUTES.accountsDetail(companyId))
-    return { success: true }
-  }
-
-  const msg = (full.error.message ?? '').toLowerCase()
-  const missingOptionalColumns =
-    msg.includes('metrics_pain') || msg.includes('mh_assessment') || msg.includes('value_proposition')
-  if (!missingOptionalColumns) return { success: false, error: full.error.message }
-
-  const fallback = await supabase.from('company_strategies').upsert(
-    {
-      company_id: companyId,
-      main_goals: payload.company_goals ?? null,
-      red_flags: payload.red_flags ?? null,
-      competitive_situation: payload.competition ?? null,
-      next_steps: payload.next_steps ?? null,
-      updated_at: new Date().toISOString(),
-    },
-    { onConflict: 'company_id' }
-  )
-  if (fallback.error) return { success: false, error: fallback.error.message }
-  revalidatePath(ROUTES.accounts)
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true }
+  return upsertCompanyStrategyImpl(companyId, payload)
 }
 
 export async function getRoadmapProjects(companyId: string): Promise<RoadmapProjectRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data } = await supabase
-    .from('company_roadmap_projects')
-    .select('*')
-    .eq('company_id', companyId)
-    .order('target_date', { ascending: true, nullsFirst: false })
-  return (data ?? []) as RoadmapProjectRow[]
+  return getRoadmapProjectsImpl(companyId)
 }
 
 export async function upsertRoadmapProject(
@@ -206,195 +105,25 @@ export async function upsertRoadmapProject(
     tags?: string | null
   }
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const row = {
-    company_id: companyId,
-    project_name: payload.project_name.trim(),
-    estimated_value: payload.estimated_value?.trim() || null,
-    status: payload.status?.trim() || null,
-    target_date: payload.target_date || null,
-    tags: payload.tags?.trim() || null,
-    updated_at: new Date().toISOString(),
-  }
-  if (payload.id) {
-    const { error } = await supabase.from('company_roadmap_projects').update(row).eq('id', payload.id)
-    if (error) return { success: false, error: error.message }
-  } else {
-    const { error } = await supabase.from('company_roadmap_projects').insert(row)
-    if (error) return { success: false, error: error.message }
-  }
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true }
+  return upsertRoadmapProjectImpl(companyId, payload)
 }
 
 export async function deleteRoadmapProject(id: string): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('company_roadmap_projects')
-    .select('company_id')
-    .eq('id', id)
-    .single()
-  const { error } = await supabase.from('company_roadmap_projects').delete().eq('id', id)
-  if (error) return { success: false, error: error.message }
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id))
-  return { success: true }
+  return deleteRoadmapProjectImpl(id)
 }
 
-function normalizeTags(tags: string | null | undefined): Set<string> {
-  if (!tags || !tags.trim()) return new Set()
-  return new Set(
-    tags
-      .split(',')
-      .map((t: string) => t.trim().toLowerCase())
-      .filter((t: string): t is string => Boolean(t))
-  )
-}
-
-function normalizeIndustry(s: string | null | undefined): string {
-  return resolveIndustryId(s)
-}
-
-/** Region/Country aus HQ-String oder Land extrahieren (normalisiert für Vergleich) */
-function normalizeRegion(s: string | null | undefined): string {
-  return (s ?? '').trim().toLowerCase().replace(/\s+/g, ' ')
-}
-
-/** Ähnliches Kundenprofil: gleiche Region (Land) oder HQ enthält Referenz-Land */
-function sizeRegionMatch(
-  companyHeadquarters: string | null | undefined,
-  refCountry: string | null | undefined
-): boolean {
-  const hq = normalizeRegion(companyHeadquarters)
-  const country = normalizeRegion(refCountry)
-  if (!hq && !country) return false
-  if (!country) return false
-  return hq.includes(country) || country.includes(hq) || hq === country
-}
-
-/** Smart-Matching: Top 3 Referenzen nach Score 0–100 (Branche 50, Themen 30, Größe/Region 20) */
 export async function getRecommendedReferences(
   projectId: string
 ): Promise<RecommendedReference[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data: project } = await supabase
-    .from('company_roadmap_projects')
-    .select('company_id, tags')
-    .eq('id', projectId)
-    .single()
-  if (!project?.company_id) return []
-  const projectCompanyId = project.company_id
-
-  const { data: company } = await supabase
-    .from('companies')
-    .select('industry, headquarters')
-    .eq('id', projectCompanyId)
-    .single()
-  const companyIndustry = normalizeIndustry(company?.industry ?? null)
-  const companyHeadquarters = company?.headquarters ?? null
-  const projectTagSet = normalizeTags(project.tags ?? null)
-
-  const { data: refRows } = await supabase
-    .from('references')
-    .select('id, title, industry, tags, country, created_at, companies(name)')
-    .is('deleted_at', null)
-
-  if (!refRows?.length) return []
-
-  const twelveMonthsAgo = new Date()
-  twelveMonthsAgo.setMonth(twelveMonthsAgo.getMonth() - 12)
-
-  const refCountry = (r: (typeof refRows)[0]) => (r.country as string | null) ?? null
-  const companyName = (r: (typeof refRows)[0]) => {
-    const c = r.companies
-    if (Array.isArray(c) && c.length > 0) return (c[0] as { name?: string }).name ?? null
-    return (c as { name?: string } | null)?.name ?? null
-  }
-
-  const scored: {
-    ref: (typeof refRows)[0]
-    matchType: RecommendedReference['matchType']
-    score: number
-    matchReasons: RecommendedReference['matchReasons']
-  }[] = []
-
-  for (const r of refRows) {
-    const refIndustry = normalizeIndustry(r.industry as string | null)
-    const refTagSet = normalizeTags(r.tags as string | null)
-    const industryMatch = !!companyIndustry && companyIndustry === refIndustry
-    const tagMatch =
-      projectTagSet.size > 0 &&
-      refTagSet.size > 0 &&
-      [...projectTagSet].some((t: string) => refTagSet.has(t))
-    const sizeRegion = sizeRegionMatch(companyHeadquarters, refCountry(r))
-
-    let points =
-      (industryMatch ? 50 : 0) + (tagMatch ? 30 : 0) + (sizeRegion ? 20 : 0)
-    if (points === 0) continue
-
-    const createdAt = r.created_at as string | null | undefined
-    if (createdAt && new Date(createdAt) >= twelveMonthsAgo) {
-      points = Math.min(100, points + 10)
-    }
-
-    let matchType: RecommendedReference['matchType'] = 'industry_only'
-    if (industryMatch && tagMatch) matchType = 'industry_and_tags'
-    else if (tagMatch) matchType = 'tags_only'
-    else if (industryMatch) matchType = 'industry_only'
-
-    scored.push({
-      ref: r,
-      matchType,
-      score: points,
-      matchReasons: { industry: industryMatch, tags: tagMatch, sizeRegion },
-    })
-  }
-
-  scored.sort((a, b) => b.score - a.score)
-  const top = scored.slice(0, 3)
-
-  return top.map(({ ref, matchType, score, matchReasons }) => ({
-    id: ref.id,
-    title: (ref.title as string) ?? '',
-    company_name: companyName(ref),
-    matchType,
-    score,
-    matchReasons,
-  }))
+  return getRecommendedReferencesImpl(projectId)
 }
 
-/** Alle Referenzen der Org für Fallback "alle Referenzen anzeigen" (z. B. Top 10) */
 export async function getReferencesForOrg(limit = 10): Promise<RecommendedReference[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data } = await supabase
-    .from('references')
-    .select('id, title, companies(name)')
-    .is('deleted_at', null)
-    .order('created_at', { ascending: false })
-    .limit(limit)
-  if (!data?.length) return []
-  const companyName = (r: (typeof data)[0]) => {
-    const c = r.companies
-    if (Array.isArray(c) && c.length > 0) return (c[0] as { name?: string }).name ?? null
-    return (c as { name?: string } | null)?.name ?? null
-  }
-  return data.map((r) => ({
-    id: r.id,
-    title: (r.title as string) ?? '',
-    company_name: companyName(r),
-    matchType: 'industry_only' as const,
-    score: 0,
-    matchReasons: { industry: false, tags: false, sizeRegion: false },
-  }))
+  return getReferencesForOrgImpl(limit)
 }
 
 export async function getStakeholders(companyId: string): Promise<StakeholderRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data } = await supabase
-    .from('stakeholders')
-    .select('*')
-    .eq('company_id', companyId)
-    .order('created_at', { ascending: true })
-  return (data ?? []) as StakeholderRow[]
+  return getStakeholdersImpl(companyId)
 }
 
 export async function createStakeholder(
@@ -413,28 +142,7 @@ export async function createStakeholder(
     sentiment?: string | null
   }
 ): Promise<{ success: boolean; stakeholder?: StakeholderRow; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data, error } = await supabase
-    .from('stakeholders')
-    .insert({
-      company_id: companyId,
-      name: payload.name.trim(),
-      title: payload.title?.trim() || null,
-      role: payload.role,
-      influence_level: payload.influence_level?.trim() || null,
-      attitude: payload.attitude?.trim() || null,
-      notes: payload.notes?.trim() || null,
-      linkedin_url: payload.linkedin_url?.trim() || null,
-      priorities_topics: payload.priorities_topics?.trim() || null,
-      last_contact_at: payload.last_contact_at || null,
-      last_interaction_at: payload.last_interaction_at || payload.last_contact_at || null,
-      sentiment: payload.sentiment?.trim() || null,
-    })
-    .select('*')
-    .single()
-  if (error) return { success: false, error: error.message }
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true, stakeholder: data as StakeholderRow }
+  return createStakeholderImpl(companyId, payload)
 }
 
 export async function updateStakeholder(
@@ -453,194 +161,35 @@ export async function updateStakeholder(
     sentiment?: string | null
   }
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('stakeholders')
-    .select('company_id')
-    .eq('id', id)
-    .single()
-  const update: Record<string, unknown> = { updated_at: new Date().toISOString() }
-  if (payload.name !== undefined) update.name = payload.name.trim()
-  if (payload.title !== undefined) update.title = payload.title?.trim() || null
-  if (payload.role !== undefined) update.role = payload.role
-  if (payload.influence_level !== undefined) update.influence_level = payload.influence_level?.trim() || null
-  if (payload.attitude !== undefined) update.attitude = payload.attitude?.trim() || null
-  if (payload.notes !== undefined) update.notes = payload.notes?.trim() || null
-  if (payload.linkedin_url !== undefined) update.linkedin_url = payload.linkedin_url?.trim() || null
-  if (payload.priorities_topics !== undefined) update.priorities_topics = payload.priorities_topics?.trim() || null
-  if (payload.last_contact_at !== undefined) update.last_contact_at = payload.last_contact_at || null
-  if (payload.last_interaction_at !== undefined) update.last_interaction_at = payload.last_interaction_at || null
-  if (payload.sentiment !== undefined) update.sentiment = payload.sentiment?.trim() || null
-  const { error } = await supabase.from('stakeholders').update(asTableUpdate<'stakeholders'>(update)).eq('id', id)
-  if (error) return { success: false, error: error.message }
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id))
-  return { success: true }
+  return updateStakeholderImpl(id, payload)
 }
 
 export async function deleteStakeholder(id: string): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('stakeholders')
-    .select('company_id')
-    .eq('id', id)
-    .single()
-  const { error } = await supabase.from('stakeholders').delete().eq('id', id)
-  if (error) return { success: false, error: error.message }
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id))
-  return { success: true }
+  return deleteStakeholderImpl(id)
 }
 
 export async function getReferencesByCompanyId(
   companyId: string
 ): Promise<CompanyRefRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data } = await supabase
-    .from('references')
-    .select('id, title, status, project_status, industry, country, created_at')
-    .eq('company_id', companyId)
-    .is('deleted_at', null)
-    .order('created_at', { ascending: false })
-  return (data ?? []) as CompanyRefRow[]
-}
-
-export type ContactPersonRow = {
-  id: string
-  company_id: string
-  first_name: string | null
-  last_name: string | null
-  email: string | null
-  phone?: string | null
-  linkedin_url?: string | null
-  role?: string | null
-  position?: string | null
-  avatar_url?: string | null
-  last_interaction_at?: string | null
-  created_at: string
-  updated_at: string | null
-}
-
-export type ExternalContactRow = {
-  id: string
-  company_id: string
-  first_name: string | null
-  last_name: string | null
-  email: string | null
-  phone: string | null
-  /** Jobtitel (z. B. CIO), nicht Miller-Heiman-Rolle */
-  role: string | null
-  buying_center_role?: StakeholderRole | null
-  last_interaction_at?: string | null
-  created_at: string
-  updated_at: string | null
+  return getReferencesByCompanyIdImpl(companyId)
 }
 
 export async function updateExternalContactBuyingCenterRole(
   id: string,
   role: StakeholderRole
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('external_contacts')
-    .select('company_id')
-    .eq('id', id)
-    .maybeSingle()
-
-  const { error } = await supabase
-    .from('external_contacts')
-    .update({ buying_center_role: role, updated_at: new Date().toISOString() })
-    .eq('id', id)
-
-  if (error) {
-    const msg = error.message ?? ''
-    if (/buying_center_role/i.test(msg)) {
-      return {
-        success: false,
-        error:
-          'Buying-Center-Rolle für Referenz-Kontakte ist in der Datenbank noch nicht verfügbar. Bitte Migration ausführen.',
-      }
-    }
-    return { success: false, error: msg }
-  }
-
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id as string))
-  return { success: true }
+  return updateExternalContactBuyingCenterRoleImpl(id, role)
 }
 
 export async function getContactsByCompanyId(companyId: string): Promise<ContactPersonRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data, error } = await supabase
-    .from('contact_persons')
-    .select('*')
-    .eq('company_id', companyId)
-    .order('created_at', { ascending: true })
-  if (error && (error.message ?? '').includes('company_id')) {
-    // DB noch ohne company_id-Spalte → keine Kontakte im Account-Detail anzeigen
-    return []
-  }
-  return (data ?? []) as ContactPersonRow[]
+  return getContactsByCompanyIdImpl(companyId)
 }
 
-/** Ein interner Kontakt pro Account als Ansprechpartner für Koordination der Referenzfreigabe (Kunde). */
 export async function setCompanyInternalReferenceApprovalContact(
   companyId: string,
   contactPersonId: string | null
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht angemeldet.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) return { success: false, error: 'Onboarding unvollständig.' }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) return { success: false, error: 'Keine Berechtigung.' }
-
-  const { data: company, error: cErr } = await supabase
-    .from('companies')
-    .select('id, organization_id')
-    .eq('id', companyId)
-    .single()
-  if (cErr || !company) return { success: false, error: 'Account nicht gefunden.' }
-  if ((company as { organization_id: string }).organization_id !== profile.organization_id) {
-    return { success: false, error: 'Keine Berechtigung.' }
-  }
-
-  if (contactPersonId) {
-    const { data: cp, error: cpErr } = await supabase
-      .from('contact_persons')
-      .select('id, company_id')
-      .eq('id', contactPersonId)
-      .single()
-    if (cpErr || !cp) return { success: false, error: 'Kontakt nicht gefunden.' }
-    if ((cp as { company_id: string }).company_id !== companyId) {
-      return { success: false, error: 'Kontakt gehört nicht zu diesem Account.' }
-    }
-  }
-
-  const { error } = await supabase
-    .from('companies')
-    .update({ internal_reference_approval_contact_id: contactPersonId })
-    .eq('id', companyId)
-    .eq('organization_id', profile.organization_id)
-
-  if (error) {
-    if ((error.message ?? '').toLowerCase().includes('internal_reference_approval')) {
-      return {
-        success: false,
-        error:
-          'Spalte internal_reference_approval_contact_id fehlt. Bitte Migration ausführen und Schema aktualisieren.',
-      }
-    }
-    return { success: false, error: error.message }
-  }
-
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true }
+  return setCompanyInternalReferenceApprovalContactImpl(companyId, contactPersonId)
 }
 
 export async function createContactPerson(
@@ -656,52 +205,7 @@ export async function createContactPerson(
     last_interaction_at?: string | null
   }
 ): Promise<{ success: boolean; contact?: ContactPersonRow; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht angemeldet.' }
-
-  // Einige Deployments verlangen organization_id auf contact_persons.
-  const [{ data: company }, { data: profile }] = await Promise.all([
-    supabase.from('companies').select('organization_id').eq('id', companyId).single(),
-    supabase.from('profiles').select('organization_id').eq('id', user.id).single(),
-  ])
-  const organization_id =
-    (company as { organization_id?: string | null } | null)?.organization_id ??
-    (profile as { organization_id?: string | null } | null)?.organization_id ??
-    null
-
-  const insertRow: Record<string, unknown> = {
-    company_id: companyId,
-    first_name: payload.first_name?.trim() || null,
-    last_name: payload.last_name?.trim() || null,
-    email: payload.email?.trim().toLowerCase() || null,
-    phone: payload.phone?.trim() || null,
-    linkedin_url: payload.linkedin_url?.trim() || null,
-    role: payload.role?.trim() || null,
-    position: payload.position?.trim() || null,
-    last_interaction_at: payload.last_interaction_at || null,
-  }
-  if (organization_id) insertRow.organization_id = organization_id
-
-  const { data, error } = await supabase
-    .from('contact_persons')
-    .insert(asTableInsert<'contact_persons'>(insertRow))
-    .select('*')
-    .single()
-  if (error) {
-    if ((error.message ?? '').includes('company_id')) {
-      return {
-        success: false,
-        error:
-          "Kontakte können noch nicht einem Account zugeordnet werden, weil die Spalte 'contact_persons.company_id' in deiner DB fehlt. Bitte Migration ausführen und Schema-Cache refreshen.",
-      }
-    }
-    return { success: false, error: error.message }
-  }
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true, contact: data as ContactPersonRow }
+  return createContactPersonImpl(companyId, payload)
 }
 
 export async function updateContactPerson(
@@ -718,161 +222,37 @@ export async function updateContactPerson(
     last_interaction_at?: string | null
   }
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('contact_persons')
-    .select('company_id')
-    .eq('id', id)
-    .single()
-  const update: Record<string, unknown> = { updated_at: new Date().toISOString() }
-  if (payload.first_name !== undefined) update.first_name = payload.first_name?.trim() || null
-  if (payload.last_name !== undefined) update.last_name = payload.last_name?.trim() || null
-  if (payload.email !== undefined) update.email = payload.email?.trim().toLowerCase() || null
-  if (payload.phone !== undefined) update.phone = payload.phone?.trim() || null
-  if (payload.linkedin_url !== undefined) update.linkedin_url = payload.linkedin_url?.trim() || null
-  if (payload.role !== undefined) update.role = payload.role?.trim() || null
-  if (payload.position !== undefined) update.position = payload.position?.trim() || null
-  if (payload.company_id !== undefined) update.company_id = payload.company_id || null
-  if (payload.last_interaction_at !== undefined) update.last_interaction_at = payload.last_interaction_at || null
-  const { error } = await supabase.from('contact_persons').update(asTableUpdate<'contact_persons'>(update)).eq('id', id)
-  if (error) return { success: false, error: error.message }
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id))
-  return { success: true }
+  return updateContactPersonImpl(id, payload)
 }
 
 export async function deleteContactPerson(id: string): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { data: row } = await supabase
-    .from('contact_persons')
-    .select('company_id')
-    .eq('id', id)
-    .single()
-  const { error } = await supabase.from('contact_persons').delete().eq('id', id)
-  if (error) return { success: false, error: error.message }
-  if (row?.company_id) revalidatePath(ROUTES.accountsDetail(row.company_id))
-  return { success: true }
-}
-
-export type AccountDealRow = {
-  id: string
-  title: string
-  volume: string | null
-  status: string
-  expiry_date: string | null
-  salesforce_opportunity_id?: string | null
-  crm_opportunity_id?: string | null
-  crm_source?: string | null
-  crm_synced_at?: string | null
-  created_at: string
-  updated_at: string | null
+  return deleteContactPersonImpl(id)
 }
 
 export async function getActiveDealsByCompanyId(companyId: string): Promise<AccountDealRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return []
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id')
-    .eq('id', user.id)
-    .single()
-  const orgId = profile?.organization_id
-  if (!orgId) return []
-
-  const { data } = await supabase
-    .from('deals')
-    .select('id, title, volume, status, expiry_date, salesforce_opportunity_id, crm_opportunity_id, crm_source, crm_synced_at, created_at, updated_at')
-    .eq('organization_id', orgId)
-    .eq('company_id', companyId)
-    .not('status', 'in', '("won","lost")')
-    .order('updated_at', { ascending: false, nullsFirst: false })
-  return (data ?? []) as AccountDealRow[]
+  return getActiveDealsByCompanyIdImpl(companyId)
 }
 
-/** Expiring Deals für einen Account (Market Signals Tab). */
-export type DealSignalRow = {
-  id: string
-  title: string
-  expiry_date: string | null
-  volume: string | null
-  incumbent_provider: string | null
-  status: string
-}
 export async function getExpiringDealsByCompanyId(
   companyId: string
 ): Promise<DealSignalRow[]> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return []
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id')
-    .eq('id', user.id)
-    .single()
-  const orgId = profile?.organization_id
-  if (!orgId) return []
-  const now = new Date()
-  const in180 = new Date(now)
-  in180.setDate(in180.getDate() + 180)
-  const { data } = await supabase
-    .from('deals')
-    .select('id, title, expiry_date, volume, incumbent_provider, status')
-    .eq('company_id', companyId)
-    .eq('organization_id', orgId)
-    .not('expiry_date', 'is', null)
-    .lte('expiry_date', in180.toISOString().slice(0, 10))
-    .order('expiry_date', { ascending: true })
-  return (data ?? []) as DealSignalRow[]
+  return getExpiringDealsByCompanyIdImpl(companyId)
 }
 
-/** Setzt nur `account_status`. Stammdaten inkl. Status: {@link updateCompany}. */
 export async function updateCompanyAccountStatus(
   companyId: string,
   account_status: CompanyAccountStatusValue | null
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { error } = await supabase
-    .from('companies')
-    .update({
-      account_status,
-      account_status_source: account_status ? 'manual' : 'crm',
-    } as { account_status: string | null; account_status_source: string | null })
-    .eq('id', companyId)
-  if (error) return { success: false, error: error.message }
-  revalidatePath(ROUTES.accounts)
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true }
+  return updateCompanyAccountStatusImpl(companyId, account_status)
 }
 
 export async function toggleCompanyFavorite(
   companyId: string,
   isFavorite: boolean
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const { error } = await supabase
-    .from('companies')
-    .update({ is_favorite: isFavorite })
-    .eq('id', companyId)
-  if (error) {
-    if ((error.message ?? '').includes('is_favorite')) {
-      return {
-        success: false,
-        error:
-          "Favoriten sind in deiner DB noch nicht aktiviert (Spalte 'companies.is_favorite' fehlt). Bitte Migration ausführen und Schema-Cache refreshen.",
-      }
-    }
-    return { success: false, error: error.message }
-  }
-  revalidatePath(ROUTES.accounts)
-  revalidatePath(ROUTES.accountsDetail(companyId))
-  return { success: true }
+  return toggleCompanyFavoriteImpl(companyId, isFavorite)
 }
+
 export async function createCompany(payload: {
   name: string
   website_url?: string | null
@@ -883,49 +263,7 @@ export async function createCompany(payload: {
   description?: string | null
   account_status?: string | null
 }): Promise<{ success: boolean; id?: string; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht eingeloggt.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) return { success: false, error: 'Onboarding unvollständig.' }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) return { success: false, error: 'Keine Berechtigung.' }
-
-  const name = payload.name.trim()
-  if (!name) return { success: false, error: 'Name ist erforderlich.' }
-
-  const { data, error } = await supabase
-    .from('companies')
-    .insert({
-      organization_id: profile.organization_id,
-      entity_kind: 'account',
-      name,
-      website_url: payload.website_url?.trim() || null,
-      industry: payload.industry?.trim() || null,
-      headquarters: payload.headquarters?.trim() || null,
-      logo_url: ensureBrandfetchDarkLogoUrl(payload.logo_url?.trim() || null),
-      employee_count: payload.employee_count ?? null,
-      description: payload.description?.trim() || null,
-      account_status: normalizeCompanyAccountStatus(payload.account_status),
-    })
-    .select('id')
-    .single()
-
-  if (error) return { success: false, error: error.message }
-
-  if (data?.id) {
-    scheduleCompanyNewsroomDiscovery(supabase, data.id, payload.website_url?.trim() || null)
-  }
-
-  revalidatePath(ROUTES.accounts)
-  return { success: true, id: data?.id }
+  return createCompanyImpl(payload)
 }
 
 export async function createPartner(payload: {
@@ -940,79 +278,7 @@ export async function createPartner(payload: {
 }): Promise<
   { success: true; id: string; accountId?: string } | { success: false; error: string }
 > {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht eingeloggt.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) return { success: false, error: 'Onboarding unvollständig.' }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) return { success: false, error: 'Keine Berechtigung.' }
-
-  const name = payload.name.trim()
-  if (!name) return { success: false, error: 'Name ist erforderlich.' }
-
-  const baseFields = {
-    organization_id: profile.organization_id,
-    name,
-    website_url: payload.website_url?.trim() || null,
-    industry: payload.industry?.trim() || null,
-    headquarters: payload.headquarters?.trim() || null,
-    logo_url: ensureBrandfetchDarkLogoUrl(payload.logo_url?.trim() || null),
-    description: payload.description?.trim() || null,
-  }
-
-  let linkedAccountId: string | null = null
-
-  if (payload.alsoCreateAccount) {
-    const { data: accountRow, error: accountError } = await supabase
-      .from('companies')
-      .insert({
-        ...baseFields,
-        entity_kind: 'account',
-        account_status: null,
-      })
-      .select('id')
-      .single()
-
-    if (accountError) return { success: false, error: accountError.message }
-    linkedAccountId = accountRow?.id ?? null
-    if (linkedAccountId) {
-      scheduleCompanyNewsroomDiscovery(supabase, linkedAccountId, baseFields.website_url)
-    }
-  }
-
-  const { data: partnerRow, error: partnerError } = await supabase
-    .from('companies')
-    .insert({
-      ...baseFields,
-      entity_kind: 'partner',
-      partner_category: payload.partner_category,
-      linked_account_id: linkedAccountId,
-      account_status: null,
-    })
-    .select('id')
-    .single()
-
-  if (partnerError) {
-    if (linkedAccountId) {
-      await supabase.from('companies').delete().eq('id', linkedAccountId)
-    }
-    return { success: false, error: partnerError.message }
-  }
-
-  revalidatePath(ROUTES.accounts)
-  return {
-    success: true,
-    id: partnerRow!.id,
-    accountId: linkedAccountId ?? undefined,
-  }
+  return createPartnerImpl(payload)
 }
 
 export async function bulkCreateCompaniesFromSheet(
@@ -1025,105 +291,7 @@ export async function bulkCreateCompaniesFromSheet(
   failedCount: number
   error?: string
 }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, createdCount: 0, skippedCount: 0, failedCount: 0, error: 'Nicht eingeloggt.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) {
-    return { success: false, createdCount: 0, skippedCount: 0, failedCount: 0, error: 'Onboarding unvollständig.' }
-  }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) {
-    return { success: false, createdCount: 0, skippedCount: 0, failedCount: 0, error: 'Keine Berechtigung.' }
-  }
-
-  const entityKind = options.entityKind === 'partner' ? 'partner' : 'account'
-
-  let workbook: XLSX.WorkBook
-  try {
-    workbook = XLSX.read(fileBuffer, { type: 'array' })
-  } catch {
-    return { success: false, createdCount: 0, skippedCount: 0, failedCount: 0, error: 'Datei konnte nicht gelesen werden.' }
-  }
-  const firstSheetName = workbook.SheetNames[0]
-  if (!firstSheetName) {
-    return { success: false, createdCount: 0, skippedCount: 0, failedCount: 0, error: 'Keine Tabelle gefunden.' }
-  }
-  const rows = XLSX.utils.sheet_to_json<Record<string, unknown>>(workbook.Sheets[firstSheetName]!, { defval: '' })
-
-  const { data: existingCompanies } = await supabase
-    .from('companies')
-    .select('name')
-    .eq('organization_id', profile.organization_id)
-    .eq('entity_kind', entityKind)
-  const existingNames = new Set((existingCompanies ?? []).map((c) => String(c.name ?? '').trim().toLowerCase()))
-
-  let createdCount = 0
-  let skippedCount = 0
-  let failedCount = 0
-
-  for (const row of rows) {
-    const parsed = parseCompaniesImportRow(row, entityKind)
-    if (!parsed) {
-      skippedCount += 1
-      continue
-    }
-    const normalizedName = parsed.name.toLowerCase()
-    if (existingNames.has(normalizedName)) {
-      skippedCount += 1
-      continue
-    }
-
-    const enriched = await enrichBulkImportRowFromBrandfetch({
-      name: parsed.name,
-      website: parsed.website,
-      industry: parsed.industry,
-      headquarters: parsed.headquarters,
-      employeeCount: parsed.employeeCount,
-    })
-
-    const partner_category = parsed.partnerCategory
-
-    const { data: inserted, error } = await supabase
-      .from('companies')
-      .insert({
-      organization_id: profile.organization_id,
-      entity_kind: entityKind,
-      name: enriched.name,
-      website_url: enriched.website.trim() || null,
-      industry: enriched.industry.trim() || null,
-      headquarters: enriched.headquarters.trim() || null,
-      employee_count:
-        enriched.employeeCount != null && Number.isFinite(enriched.employeeCount)
-          ? enriched.employeeCount
-          : null,
-      logo_url: ensureBrandfetchDarkLogoUrl(enriched.logo_url),
-      description: enriched.description?.trim() || null,
-      account_status: null,
-      ...(entityKind === 'partner' ? { partner_category } : {}),
-    })
-      .select('id')
-      .maybeSingle()
-    if (error) {
-      failedCount += 1
-      continue
-    }
-    if (inserted?.id && entityKind === 'account') {
-      scheduleCompanyNewsroomDiscovery(supabase, inserted.id, enriched.website.trim() || null)
-    }
-    existingNames.add(normalizedName)
-    createdCount += 1
-  }
-
-  revalidatePath(ROUTES.accounts)
-  return { success: true, createdCount, skippedCount, failedCount }
+  return bulkCreateCompaniesFromSheetImpl(fileBuffer, options)
 }
 
 export async function updateCompany(payload: {
@@ -1137,259 +305,23 @@ export async function updateCompany(payload: {
   description?: string | null
   account_status?: string | null
 }): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht eingeloggt.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) return { success: false, error: 'Onboarding unvollständig.' }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) return { success: false, error: 'Keine Berechtigung.' }
-
-  const name = payload.name.trim()
-  if (!name) return { success: false, error: 'Name ist erforderlich.' }
-
-  const account_status = normalizeCompanyAccountStatus(payload.account_status)
-
-  const { data: row, error: fetchError } = await supabase
-    .from('companies')
-    .select('id, organization_id, website_url')
-    .eq('id', payload.id)
-    .single()
-
-  if (fetchError || !row) return { success: false, error: 'Account nicht gefunden.' }
-  if (row.organization_id !== profile.organization_id) {
-    return { success: false, error: 'Keine Berechtigung.' }
-  }
-
-  const nextWebsite = payload.website_url?.trim() || null
-  const prevWebsite = (row as { website_url?: string | null }).website_url ?? null
-
-  const { error } = await supabase
-    .from('companies')
-    .update({
-      name,
-      website_url: nextWebsite,
-      industry: payload.industry?.trim() || null,
-      headquarters: payload.headquarters?.trim() || null,
-      logo_url: ensureBrandfetchDarkLogoUrl(payload.logo_url?.trim() || null),
-      employee_count: payload.employee_count ?? null,
-      description: payload.description?.trim() || null,
-      account_status,
-    })
-    .eq('id', payload.id)
-    .eq('organization_id', profile.organization_id)
-
-  if (error) return { success: false, error: error.message }
-
-  if (nextWebsite && nextWebsite !== prevWebsite) {
-    void discoverAndSaveCompanyNewsrooms(supabase, payload.id, {
-      websiteUrl: nextWebsite,
-      force: true,
-    }).catch((err) => {
-      console.error('[newsroom-discover]', payload.id, err instanceof Error ? err.message : err)
-    })
-  }
-
-  revalidatePath(ROUTES.accounts)
-  revalidatePath(ROUTES.accountsDetail(payload.id))
-  return { success: true }
+  return updateCompanyImpl(payload)
 }
 
 export async function deleteCompanyWithData(
   companyId: string
 ): Promise<{ success: boolean; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser()
-  if (!user) return { success: false, error: 'Nicht angemeldet.' }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('organization_id, system_role, function_role')
-    .eq('id', user.id)
-    .single()
-  if (!profile?.organization_id) return { success: false, error: 'Onboarding unvollständig.' }
-  const { systemRole, functionRole } = parseProfileRoles(profile)
-  if (profileIsSalesRestricted(systemRole, functionRole)) {
-    return { success: false, error: 'Keine Berechtigung.' }
-  }
-
-  const { data: company, error: companyErr } = await supabase
-    .from('companies')
-    .select('id, organization_id')
-    .eq('id', companyId)
-    .single()
-  if (companyErr || !company) return { success: false, error: 'Account nicht gefunden.' }
-  if ((company as { organization_id: string }).organization_id !== profile.organization_id) {
-    return { success: false, error: 'Keine Berechtigung.' }
-  }
-
-  // Optional: weitere abhängige Daten explizit löschen, falls kein ON DELETE CASCADE konfiguriert ist.
-  // Referenzen
-  await supabase.from('references').delete().eq('company_id', companyId)
-  // Deals
-  await supabase.from('deals').delete().eq('company_id', companyId)
-  // Strategy
-  await supabase.from('company_strategies').delete().eq('company_id', companyId)
-  // Roadmap-Projekte
-  await supabase.from('company_roadmap_projects').delete().eq('company_id', companyId)
-  // Stakeholder
-  await supabase.from('stakeholders').delete().eq('company_id', companyId)
-
-  const { error } = await supabase
-    .from('companies')
-    .delete()
-    .eq('id', companyId)
-    .eq('organization_id', profile.organization_id)
-  if (error) return { success: false, error: error.message }
-
-  revalidatePath(ROUTES.accounts)
-  return { success: true }
+  return deleteCompanyWithDataImpl(companyId)
 }
 
-/** Smart Match für Account: Referenzen aus der Org, die zu diesem Kunden passen (Branche/Herausforderungen). */
 export async function getRecommendedReferencesForAccount(
   companyId: string
 ): Promise<RecommendedReference[]> {
-  const supabase = await createServerSupabaseClient()
-  const { data: company } = await supabase
-    .from('companies')
-    .select('industry, headquarters')
-    .eq('id', companyId)
-    .single()
-  const { data: strategy } = await supabase
-    .from('company_strategies')
-    // main_goals -> company_goals
-    .select('company_goals:main_goals')
-    .eq('company_id', companyId)
-    .maybeSingle()
-  const companyIndustry = normalizeIndustry(company?.industry ?? null)
-  const goalsTags: string[] = (strategy?.company_goals ?? '')
-    .split(/[\s,;]+/)
-    .map((t: string) => t.trim().toLowerCase())
-    .filter((t: string): t is string => Boolean(t))
-  const projectTagSet = new Set<string>(goalsTags)
-  const companyHeadquarters = company?.headquarters ?? null
-
-  const { data: refRows } = await supabase
-    .from('references')
-    .select('id, title, industry, tags, country, created_at, companies(name)')
-    .is('deleted_at', null)
-    .neq('company_id', companyId)
-
-  if (!refRows?.length) return []
-  const twelveMonthsAgo = new Date()
-  twelveMonthsAgo.setMonth(twelveMonthsAgo.getMonth() - 12)
-  const refCountry = (r: (typeof refRows)[0]) => (r.country as string | null) ?? null
-  const companyName = (r: (typeof refRows)[0]) => {
-    const c = r.companies
-    if (Array.isArray(c) && c.length > 0) return (c[0] as { name?: string }).name ?? null
-    return (c as { name?: string } | null)?.name ?? null
-  }
-  const scored: {
-    ref: (typeof refRows)[0]
-    matchType: RecommendedReference['matchType']
-    score: number
-    matchReasons: RecommendedReference['matchReasons']
-  }[] = []
-  for (const r of refRows) {
-    const refIndustry = normalizeIndustry(r.industry as string | null)
-    const refTagSet = normalizeTags(r.tags as string | null)
-    const industryMatch = !!companyIndustry && companyIndustry === refIndustry
-    const tagMatch =
-      projectTagSet.size > 0 &&
-      refTagSet.size > 0 &&
-      [...projectTagSet].some((t: string) => refTagSet.has(t))
-    const sizeRegion = sizeRegionMatch(companyHeadquarters, refCountry(r))
-    let points =
-      (industryMatch ? 50 : 0) + (tagMatch ? 30 : 0) + (sizeRegion ? 20 : 0)
-    if (points === 0) continue
-    const createdAt = r.created_at as string | null | undefined
-    if (createdAt && new Date(createdAt) >= twelveMonthsAgo) points = Math.min(100, points + 10)
-    let matchType: RecommendedReference['matchType'] = 'industry_only'
-    if (industryMatch && tagMatch) matchType = 'industry_and_tags'
-    else if (tagMatch) matchType = 'tags_only'
-    scored.push({
-      ref: r,
-      matchType,
-      score: points,
-      matchReasons: { industry: industryMatch, tags: tagMatch, sizeRegion },
-    })
-  }
-  scored.sort((a, b) => b.score - a.score)
-  return scored.slice(0, 5).map(({ ref, matchType, score, matchReasons }) => ({
-    id: ref.id,
-    title: (ref.title as string) ?? '',
-    company_name: companyName(ref),
-    matchType,
-    score,
-    matchReasons,
-  }))
+  return getRecommendedReferencesForAccountImpl(companyId)
 }
 
-/** One-Pager HTML für Druck/PDF: Strategy + Stakeholder-Prioritäten + Referenzen. */
 export async function generateOnePagerHtml(
   companyId: string
 ): Promise<{ success: boolean; html?: string; error?: string }> {
-  const supabase = await createServerSupabaseClient()
-  const [
-    { data: company },
-    { data: strategy },
-    { data: stakeholders },
-    refs,
-  ] = await Promise.all([
-    supabase.from('companies').select('name, industry').eq('id', companyId).single(),
-    // main_goals -> company_goals; value_proposition existiert evtl. nicht in allen Deployments, daher hier nicht selektieren
-    supabase.from('company_strategies').select('company_goals:main_goals, red_flags, next_steps').eq('company_id', companyId).maybeSingle(),
-    supabase.from('stakeholders').select('name, title, role, priorities_topics').eq('company_id', companyId),
-    getReferencesByCompanyId(companyId),
-  ])
-  if (!company) return { success: false, error: 'Unternehmen nicht gefunden.' }
-  const goals = strategy?.company_goals ?? ''
-  const challenges = strategy?.red_flags ?? ''
-  const valueProp = (strategy as { value_proposition?: string | null } | null)?.value_proposition ?? ''
-  const nextSteps = strategy?.next_steps ?? ''
-  type StakeholderData = { name: string; title?: string | null; role: string | null; priorities_topics?: string | null }
-  const stakeholderList: StakeholderData[] = (stakeholders ?? []).map((s) => ({
-    name: s.name,
-    title: s.title,
-    role: s.role,
-    priorities_topics: s.priorities_topics,
-  }))
-  const execSummary = stakeholderList.map((s: StakeholderData) => `${s.name}${s.title ? ` (${s.title})` : ''}: ${(s.priorities_topics ?? '').trim() || '—'}`).join('\n')
-  const refList = refs.map((r: CompanyRefRow) => r.title).join(', ') || '—'
-  const html = `<!DOCTYPE html><html><head><meta charset="utf-8"/><title>One-Pager ${escapeHtml(company.name)}</title><style>body{font-family:system-ui,sans-serif;max-width:800px;margin:2rem auto;padding:0 1rem;line-height:1.5;} h1{font-size:1.5rem;} h2{font-size:1.1rem;margin-top:1.5rem;} ul{margin:0.25rem 0;} .meta{color:#666;font-size:0.9rem;}</style></head><body>
-<h1>${escapeHtml(company.name)}</h1>
-<p class="meta">${escapeHtml(formatIndustryDisplay(company.industry) || '')}</p>
-<h2>Unternehmensziele</h2>
-<p>${escapeHtml(goals) || '—'}</p>
-<h2>Value Proposition (Warum gewinnen wir hier?)</h2>
-<p>${escapeHtml(valueProp) || '—'}</p>
-<h2>Herausforderungen / Red Flags</h2>
-<p>${escapeHtml(challenges) || '—'}</p>
-<h2>Entscheider & Prioritäten</h2>
-<pre style="white-space:pre-wrap;font-size:0.9rem;">${escapeHtml(execSummary) || '—'}</pre>
-<h2>Nächste Schritte</h2>
-<p>${escapeHtml(nextSteps) || '—'}</p>
-<h2>Referenzen / Proof Points</h2>
-<p>${escapeHtml(refList)}</p>
-<p class="meta" style="margin-top:2rem;">Erstellt mit Client Intelligence · ${new Date().toLocaleDateString('de-DE')}</p>
-</body></html>`
-  return { success: true, html }
+  return generateOnePagerHtmlImpl(companyId)
 }
-function escapeHtml(s: string): string {
-  return s
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-}
-
